git-svn-id: https://local.ugene.unipro.ru/svn/ugene/trunk@1297 7ca3a678-163c-e011-b84f-001e8c0267ff

+++ src/corelibs/U2View/src/ov_sequence/ADVSingleSequenceWidget.cpp
@@ -472,7 +472,7 @@ bool ADVSingleSequenceWidget::eventFilter(QObject* o, QEvent* e) {
     return false;
 }
 
-void ADVSingleSequenceWidget::sl_onLocalCenteringRequest(int pos) {
+void ADVSingleSequenceWidget::sl_onLocalCenteringRequest(qint64 pos) {
     detView->setCenterPos(pos);
 }
